refact : profiles view spectacular 코드 분리
데코레이터 패턴 적용
